update typing for mypy 1.11; pin plugin to <1.11
(1.4 specific notes)

The legacy mypy plugin is no longer fully functional with the latest series
of mypy 1.11.0, as changes in the mypy interpreter are no longer compatible
with the approach used by the plugin.  If code is dependent on the legacy
mypy plugin with sqlalchemy2-stubs, it's recommended to pin mypy to be
below the 1.11.0 series.    Seek upgrading to the 2.0 series of SQLAlchemy
and migrating to the modern type annotations.

Change-Id: Ib8fef93ede588430dc0f7ed44ef887649a415821
cherry picked from commit 0741fe45f62f89845b15d7faea209b26a652a174)

Change-Id: I61f50539cda851b98178060410fedaa70971d01a