correct mis-cherry-picked commit for #11449 in 2.0.31
The main review for #11449 at Ie8f0e8d9bb7958baac33c7c2231e4afae15cf5b1
had three revisions but the cherry pick to 2.0 somehow missed the
second two cherry picks.  it's not clear if this was intentional.
however, as we need to continue on with correcting this behavior
it seems like we really should have the full "main" behavior in
the release.

Fixes: #11449

Change-Id: I1e79c4e0c4843268b2fce7fc2046900bd7b48f00