lookup "secondary" directly, dont use eval()
The :paramref:`_orm.relationship.secondary` parameter no longer uses Python
``eval()`` to evaluate the given string.   This parameter when passed a
string should resolve to a table name that's present in the local
:class:`.MetaData` collection only, and never needs to be any kind of
Python expression otherwise.  To use a real deferred callable based on a
name that may not be locally present yet, use a lambda instead.

Fixes: #10564
Change-Id: I9bb5a2ea17c7efac88df1470d109970cfb4c4874