apply underscores to ORM class and def names
criteria used here is:

* The class or def should definitely not be used directly by
  a third party
* The class would never be the subject of an `isinstance()` check
* The class is not exposed as the type of a return value for a public
  function

A sweep through class and function names in the ORM renames many classes
and functions that have no intent of public visibility to be underscored.
This is to reduce ambiguity as to which APIs are intended to be targeted by
third party applications and extensions.  Third parties are encouraged to
propose new public APIs in Discussions to the extent they are needed to
replace those that have been clarified as private.

Fixes: #10497
Change-Id: I9900e759be8510e01bba2c25984b9f30dc1fa9c0