Add type annotations for postgresql.array()
The type argument of array is inferred from the 'clauses' argument of
the constructor; hence array([1, 2]) has type array[int]. We explicitly
define the 'type_' parameter of __init__() as it helps type inference
when using this argument, e.g. array([], type_=CHAR) is inferred as
array[str]. Consistency between 'clauses' and 'type_' is also ensured.