Cast empty PostgreSQL ARRAY from the type specified to array()
When using `postgresql.array()` with an empty list as `clauses` and a
`type_` argument specified, we now cast the resulting `ARRAY[]` SQL
expression using that type information thus removing the need for an
explicit cast on user side.

We need to add the 'type' attribute to the cache key of
`postgresql.array` in order to distinguish statements produced with an
empty list clause and with or without a `type_` argument.