WIP: Retrieve "included" columns in unique/primary key constraint from the index
Replace the previous approach relying on a regex on the result of
pg_get_constraintdef() with a subquery on pg_attribute joined with
pg_index using the 'indkey' column (and excluding 'indnkeyatts').

In tests, in order to make sure the order is okay, we add more columns
to the reflected table and create the constraint with an order different
from table attributes numbering.

TODO:
- add condition on server_version for the 'include' subquery