Reflect index's column operator class on PostgreSQL
Fill the `postgresql_ops` key of PostgreSQL's `dialect_options` returned
by get_multi_indexes() with a mapping from column names to the operator
class, if it's not the default for respective data type.

As we need to join on ``pg_catalog.pg_opclass``, the table definition is
added to ``postgresql.pg_catalog``.

Fixes #8664.